+ fix na model_id ? model_id ma zostać float'em?
git-svn-id: svn+ssh://gamma/home/samba/koturj/svn/inz@111 e70bc5d4-e669-4a26-a90f-bfde2a246bde

+++ src/mem/db/planet_row.h
@@ -27,7 +27,7 @@ namespace MEM
 			float xvel;
 			float yvel;
 			float zvel;
-			uint8_t model_id;
+			float model_id;
 
 		private:
 			static const std::string save_string;
